refactor: frontend/cs subpackages to match new interfaces
